Changed rmq publish to use seperate connection
It didn't make sense that the user had to run the event loop just to
get e.g. a broadcast message, just to be sent out and reach the exchange.
Now we use a seperate blocking connection to achieve this.  This way
once the call returns the message has been delivered to the exchange
for sure.