Added await() to wait in general
This is because if you have a consumer you need to be able to away
in general for incoming messages i.e. you don't have a specific
future that you're waiting for to complete.  Internally (for RMQ)
this just runs the event loop.

Made Task and RPC messages mandatory meaning if they can't be delivered
to a queue immediately the message will fail.

Added a couple of examples based on the RabbitMQ website.