Fixed bug in rmq RCP call with 'nested' futures
If a RCP call returned a future that itself resolved to a future (which
could itself resolve to a future, etc) it would not work past the first
level of futures (feels like some inception shiz here).

So, this has now been remedied to allow infinitely many layers of
futures.