Fixed nasty tornado future related bug
Callbacks were being schedule on a different event loop (different from
the one used by the current communicator) upon certain futures being
resolved because tornado uses the 'current' (i.e.
tornado.ioloop.IOLoop.current()) loop to schedule callbacks.  This is
normally fine except for if the future is resolved on another thread
which is the case with the RmqThreadCommunicator.