Changed RMQ broadcast to only bind to queue when needed
Before the RMQ broacast communicator would subscribe itself to the
broadcast queue on construction but, by default, have no listeners on
that queue.  This would mean that it would intercept messages even if
there was no one to pick them up.  No it binds itself to the queue only
if it has listeners.