Fixed broadcast bug in RmqCommunicator
The communicator was creating a broadcast queue when it was created with
a TTL meaning that the queue would be deleted if there were no consumers
(after the initial consumer had been added and cancelled) and the TTL
elapsed.  This mean meant that when we came to add another broadcast
susbscriber it would try and consume on a queue that didn't exist
knocking over the entire connection.

To fix this I've just removed the TTL on the broadcast queue (although
it remains exclusive meaning that it will be deleted when the consumer
disconnects).  This should work well however it could have some minor
performance/functionality implications:

1.  If the connection drops the automatic reconnect of RobustQueue will
kick in, redeclare the queue, rebind and reconsume.  Fine, but the queue
will have to be created anew, unlike before where if the reconnect
happened before the TTL was up it would have still been there.

2.  If the connection drops the queue will be erased and any incoming
messages from this period until we reconnect/rebind/reconsume will be
lost.  Hopefully this isn't catastrophic, but again it would have been
nice to keep the queue alive and have the messages sit there until we
could reconnect.