Update `pyyaml` to prevent arbitrary code execution
Before `pyyaml==5.1` the `yaml.load` function was vulnerable to
arbitrary code execution, because it loaded the full set of YAML. There
was an alternative `safe_load` but this was not the default and could
only load a sub set of the markup language. The new version of pyyaml
deprecates the old vulnerable code and provides the `FullLoader` that
can load the full set without being vulnerable. The `safe_load` is
syntactic sugar for `load` with `Loader=FullLoader`.