Removed dependence on topika and moved to py>=3.5
Dropping python 2.7 support which also means we can move away from
topika which was a packport of aio_pika.  This has also meant that we
can drop the direct dependence on tornado which was only being used for
its backwards compatible event loop.  This has been replaced with
asyncio.  Otherwise the API should remain the same both in terms of
functionality and interface.