Added close method and corresponding exception
These are added on the communicator interface to be able to close and to
have an exception to check against if the communicator is closed.

Test added.

Exceptions have been moved to their own file but references have been
left so that they can still be imported using kiwipy.<exception> not
breaking the API.

Made sure to clean up references in RmqThreadCommunicator.close()