Added custom task queues to the rmq thread communicator (#43)
Summary:

  * Added custom task queues via comm.task_queue('my_new_queue')
  * Supported in both coroutine and thread RMQ communicators
  * new dependency: pytray. Responsible for translation of thread->async calls.
  * Corresponding tests added
  * Interface *hasn't*  been exposed at kiwipy.Communicator level

Details:

For now, this isn't exposed at the `kiwipy.Communicator` level.  This
will be added later but for now it works well with kiwi and it proved a
good testing ground for developing the feature.

It works like this:

  task_queue = comm.task_queue('my_task_queue')
  task_queue.add_task_subscriber(lambda _, task: print(task))
  task_queue.task_send('Do this!')

Essentially every communicator has a default queue (which is the one
referred to by doing comm.add_task_subscriber/task_send) but users are
free to create additional ones using the above syntax.

Async versions also exist and can be iterated like this:

await for task in task_queue:
  task.process().set_result('done')

The threaded version also supports this without the await syntax.

PyTray provides a `LoopScheduler` which does all the negotiation between
the thread running the asyncio event loop and the coroutines themselves.

task_get raises `kiwipy.QueueEmpty` exception if empty.