`Communicator.remove_task_subscriber` takes identifier instead of subscriber (#44)
Removing RMQ task subscribers was failing beacuse of the way we used to
wrap the function to get it to play nice in the event loop. Then when
the user would give us the function to unsubscribe we couldn't find it
because, in fact, we held a wrapped function, not the original.

For the sake of simplicity the API has been changed to that of broadcast
and RPC where the user either supplies or is given a unique identifier
which can later be used to remove the subscriber.

This API change will only cause problems for clients that were removing
task subscribers manually (rather than just closing the communicator).
I believe AiiDA just closes the communicator and so should not require
any changes.

Related bug fixed: exceptions in tasks were not propagating the error
message to the `RemoteException` because of a bug in `IncomingMessage`.