Feature: add global `kiwipy.connect` factory (#45)
This uses the URI to determine which backend to use (we only have the
RMQ one anyway) and passes any other arguments on to the corresponding
communicator. This makes it easier for someone to create a communicator
and also promotes abstraction by separating the client from the concrete
class that the URI will resolve to.