Make the `RmqThreadCommunicator` usable in Jupyter notebooks
The `RmqThreadCommunicator` could not be used in Jupyter notebooks as it
was creating an event loop in addition to the one used by Jupyter,
causing `asyncio` to raise an exception about the event loop already
being in a running state.

To fix this, we let the `RmqThreadCommunicator` create its event loop in
a separate thread, by having the `RmqCommunicator.connect` method simply
pipe the arguments through to the class which will spin up the event loop
when `start` is called and at that point connect to RMQ. This way, the
communicators loop is being done on the separate thread and it does not
interfere with the Jupyter's loop.

The change necessitated changing the constructor signature for both the
`RmqCommunicator` as well as the `RmqThreadCommunicator`, however the
respective `connect` method signatures remain unchanged.

The fix is tested through an actual Jupyter notebook that is run through
the pytest plugin `pytest-notebook` which is a new `dev` dependency.

In addition, a convenience function `async_connect` is added, which
creates a `RmqCommunicator` mirroring the threaded version. It is now
also possible to use an environmental variable to change the
RabbitMQ URI because, e.g., gitlab CI doesn't use 'localhost'.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>