Fixed jupyter problem with RmqThreadCommunicator
Changed connect() so that now it pipes the arguments through to the
class which will spin up the event loop in a separate thread when
start() is called and at that point connect to RMQ.  This way, all the
loop stuff is being done on the separate thread the jupyter is happy.

This necessitated changing the constructor signature for both
RmqCommunicator and RmqThreadCommunicator, however the respective
connect() function signatures remain unchanged.

Added an async_connect() that creates an RmqCommunicator mirroring the
threaded version.  Updated the tests to use this.

Also made it possible to use an environmental variable to change the
RabbitMQ URI because, e.g., gitlab CI doesn't use 'localhost'.

Added jupyter notebook test - this used to fail because of the loop
issues using RmqThreadCommunicator.  It needs to be ran interactively
though.