RmqCommunicator now uses lazy queue creation
The communicator is passed a connection but doesn't create the relevant
queues until the user carried out an action that needs one of them to
eixst such as sending a message or requesting a publisher or subscriber.
Because of this laziness, the RmqThreadCommunicator (and other clients)
don't see the change.  The only aspect that has had to change is the
getters for publishe/subscriber and queue which have now become
coroutines so they can do the relevant queue creation the first time.