Increase the default TTL for task messages
The default was set to a week which was definitely to short for the
typical use cases of `kiwipy` in high-performance computing. It is not
unlikely that processes are not worked on for more than a week, for
example during a cluster outage. Since having too long a TTL should not
affect the application too negatively, changing it to 120 days should
provide a safer default and prevent unnecessary loss of processes.