Remove `prospector` and run `pylint` directly
Also run `pylint` and `yapf` directly from revision through `pre-commit`.
Note that we still need to install an explicit version of `pylint`
through the setup as we define `language=system`. If we don't, then the
linter will complain about third-party dependencies not being importable.

A few additional pre-commit hooks are added, for example to
automatically add the encoding pragma, and some of the style rules for
`yapf` have been adapted, to match those of all other projects within
the `aiidateam` organization.