Dependencies: update requirement `pytest-asyncio~=0.12` (#82)
Version `0.11.0` of `pytest-asyncio` introduced a change that breaks our
test because the `event_loop` fixture will no longer be run as fast as
possible causing tests that use it to start a different event loop. This
would throw an error saying that the future was attached to a different
event loop. This is fixed in `pytest-asyncio==0.12.0`.