Fixed warning caused by excepted task and no reply
Python would warn when a task is completed that was using no_reply=True
and had an exception.  This is because the internal future associated
with the task never had its outcome checked in this case.  We now check
it and simply ignore in the case of no_reply=True.  This is OK because
the only way this could arise is:
1. The used called .processing() in a context in which case the exception
is raise to the user anyway.
2. The user called .process(), got given a future, and then set the
exception on the future themselves, in which case they know about it.