Revert "Increase the default TTL for task messages (#59)"
This reverts commit 42fb6f893c4bd150757e0cd0cef8c6b195b00007.

Changing the TTL of messages of the task queue was necessary because the
original value of 7 days was not sufficient for the AiiDA use case.
However, when this parameter is changed, one cannot connect to an
existing queue that was configured with the old value. The only way is
to destroy the queue and create a new one. Since this is not feasible
through a Python API, let alone in an automated fashion, this will
render all upgraded AiiDA installations completely broken. The only
solution is to temporarily revert it to the original value.