Dependencies: Update requirement for `aio-pika~=8.0`
The new major versions `7.0` and `8.0` come with important fixes that
could improve the stability of robust connections, so the minimum
required version is upped to `~=8.2`.

The requirement for `pamqp` also had to be upgraded, since this is
required by `aiormq` which is a direct dependency of `aio-pika`.

The newer version required a number of changes in the code:

* `pamqp`:

  - `specification.Basic.Ack` renamed to `commands.Basic.Ack`

* `aio-pika`:

  - `types.CloseCallbackType` renamed to `abc.ConnectionCloseCallback`
  - `Connection.add_close_callback` was replaced by
    `Connection.close_callbacks.add`
  - The `ack` and `nack` methods were made asynchronous
  - `Connection.connection` moved to `Connection.transport.connection`