Restore intended interface of `RmqThreadCommunicator` (#124)
In commit 3806a9ecb58271526347d92d9621f560b4c61c97 the requirement of the
`aio-pika` dependency was updated. In adjusting the code to the changes
in their API, certain methods of the `RmqThreadCommunicator` were made
asynchronous. However, the whole point of the `RmqThreadCommunicator` is
to shield the consumer from any async code.

The changes in the interface and the tests are reverted and the
implementation of `RmqThreadCommunicator` is updated to match the inter-
face of the newer `aio-pika` version.

Doing so actually revealed a bug in `pytray` causing the unit test
`test_queue_get_next` to hang forever while waiting for the future to be
set. The `pytray.LoopScheduler` was propagating a future the wrong way
around: instead of the result from a thread future being passed to the
asyncio future, it was doing the opposite causing the future to never be
resolved. The bug is fixed in `pytray==0.3.4`.